requirements.txt
mysql-connector-python
faker
python-dotenv